added exception handling for user and set up controllerAdvice

+++ src/main/java/cires/bemodule/exceptions/UserNotFoundException.java
@@ -2,7 +2,7 @@ package cires.bemodule.exceptions;
 
 public class UserNotFoundException extends RuntimeException{
 
-        public UserNotFoundException(Integer id) {
+        public UserNotFoundException(Long id) {
             super("User not found with id: " + id);
         }
 }
